feat: fix linting
