+++ src/app/(home)/page.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import HomePage from './home-page'
 
 export default async function Page() {
