+++ src/app/article-details/page.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import ArticleDetails from './article-details'
 
 export default async function Page() {
