+++ src/app/buy-article/page.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import BuyArticle from './buy-article'
 
 export default async function Page() {
