+++ src/app/forgot-password/page.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import ForgotPass from './forgot-password'
 
 export default async function Page() {
