+++ src/app/reader-register/page.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import ReaderRegis from './reader-register'
 
 export default async function Page() {
