+++ src/app/verification/[token]/page.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import Verification from './verification'
 
 export default async function Page() {
