+++ src/app/writer-dashboard/article-list/page.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import ArticleList from './article-list'
 
 export default async function Page() {
