+++ src/app/writer-dashboard/create-article/page.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import CreateArticle from './create-article'
 
 export default async function Page() {
