+++ src/app/writer-dashboard/edit-article/[id]/page.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import EditArticle from './edit-article'
 
 export default async function Page() {
