+++ src/app/writer-dashboard/page.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import WriterDashboard from './writer-dashboard'
 
 export default async function Page() {
