+++ src/app/writer-register/page.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import WriterRegis from './writer-register'
 
 export default async function Page() {
