--- src/pages/api/hello.ts
@@ -1,3 +1,0 @@
-export default function handler(req: any, res: any) {
-  res.status(200).json({ name: 'John Doe' })
-}
