+++ src/utils/fetchJson.ts
@@ -1,26 +1,3 @@
-export default async function fetchJson<JSON = unknown>(
-  input: RequestInfo,
-  init?: RequestInit,
-): Promise<JSON> {
-  const response = await fetch(input, init);
-
-  // if the server replies, there's always some data in json
-  // if there's a network error, it will throw at the previous line
-  const data = await response.json();
-
-  // response.ok is true when res.status is 2xx
-  // https://developer.mozilla.org/en-US/docs/Web/API/Response/ok
-  if (response.ok) {
-    return data;
-  }
-
-  throw new FetchError({
-    message: response.statusText,
-    response,
-    data,
-  });
-}
-
 export class FetchError extends Error {
   response: Response;
 
@@ -52,4 +29,31 @@ export class FetchError extends Error {
     this.data = data ?? { message: message };
   }
 }
+
+export default async function fetchJson<JSON = unknown>(
+  input: RequestInfo,
+  init?: RequestInit,
+): Promise<JSON> {
+  const response = await fetch(input, init);
+
+  // if the server replies, there's always some data in json
+  // if there's a network error, it will throw at the previous line
+  // eslint-disable-next-line
+  const data = await response.json();
+
+  // response.ok is true when res.status is 2xx
+  // https://developer.mozilla.org/en-US/docs/Web/API/Response/ok
+  if (response.ok) {
+    // eslint-disable-next-line
+    return data;
+  }
+
+  throw new FetchError({
+    message: response.statusText,
+    response,
+    // eslint-disable-next-line
+    data,
+  });
+}
+
